+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Avoid incorrect circular dependency errors from `:as-alias` by working around clj-depend bug.
+- Fix inline-def to work with defs with metas.
 
 ## 2026.02.20-16.08.58